4th edits?
the colors are in the proper spots (adding 90 to all degree values to bring the top back to red), the map will move based on pressing the up arrow, but won't continue moving. it is one movement at the specified value for every press. need to change that to a constant movement.