5th edits
this is all kinds of messed up. the methods at the bottom of the page are meant to find the index of the max height value and min height value of the generated terrain. right now, they are giving the same value as each other AND giving the same value every time it runs. need to fix this and figure out how to apply that to the h-value of the HSB to scale the color so that red is always at the top and blue is always at the bottom