6th edits
still trying to figure out everything from 5th edits. right now i am trying to link the z variable generated within the shape for loops of the draw method to the "float value = z" portion of the find value functions at the bottom of the file by making z a variable held universally outside of methods. right now, z gets defined to the value i want while its in the for loop, but then it changes back to 0 because it had to be a static variable. the 0 gets carried to the value functions, which messes with their outputs